move slot behavior to module
